V0.006
Phew its been a while but I realized during my path finding research that since I'm working of a grid instead of a graph, the number of possible avenues to go through would take to long. So this next leg is a 3 step phase to getting a working pathfinder. A bit of a set back, but it ain't SO bad.