V0.011
Now that pathfinding is working, I have now taken a completely different direction and looked at how I am going to implement the ui of the project. I started with a simple setup file that takes in a few essential variables for the interpreter to build from. I will have to implement a way to  build maps from a text file to accomodate the new format. Surface level thoughts on what needs to be done next. It feels like I'm playing suduko and filling in the numbers at random.